Do not display spouse info if No Spouse is set
The 'No Spouse' tick box being set on a profile seems to result in Spouses = []
and DataStatus.Spouse == 'blank'. This information is now used to prevent a
spouse box from being drawn with ? and ? - ? for name and lifespan.  This was
done by creating a NoSpouse object that mimics a Person with a minimal methods,
which is used as the spouse in a Couple object.

This change also removed Siblings from the request fields in the API call (that
information was never used) and adding the DataStatus field.