Improve behaviour in ancestor tree with married first cousins
Previously only a single branch would be displayed. Now all duplicated branches are
displayed, but they have a shared fate. i.e. If one is closed or expanded, all
duplicated ancestor branches are closed or expanded. Unfortunately there are still
issues with inconsistent behaviour in duplicated descendent branches.

This change also fixes a bug in the display of descendent nodes where male and
female partners were not placed consistently in the same position in a couple.