Add a people cache
WikiTreeAPI.getPerson() was changed to store newly created Person objects
in a local cache and first look there when any new request is made.
A WikiTreeAPI.makePerson() was introduced to do the same whenever a
Person object needs to be created from a JSON data collection.

While this does reduce API calls to some degree in some of the current
view implementaions, it will make a significant difference in upcoming
changes for a couples view.