Corrected name construction in Ancestor Web
Provision is made for absence of some fields in private profiles, as well
as for profiles that put all middel names in FirstName.

This logic should really be in common code, but we will do that later.

This also fixes a bug in getShortName of WikiTreeDynamicViewer.

A number of other changes in WebsView is due prettier.