Move people cache completely into the couples view
In this version there is only ever one person object for a profile
and a person object does not contain other person objects, only
their ids. To obtain such a related object, it has to be fetched
from the cache.

Any object present in the cache gets updated in place. All "copies"
of it in the tree will therefore get updated at the same time.