Match preferred name on word boundaries in first names
When deciding whether or not to include the preferred name in the
constructed name, search for it as a whole word in the other name
parts rather than any substring. Therefore, if the preferred name
appears as one of the other names, it will be omitted, but if it
is only part of any of the other names (or does not appear at all,
it will be included.