Add Ancestor Lines Explorer and upgrade to D3 v7
Ancestor Line Explorer allows for the exploration of specific lines
in an ancestor tree. The ancestor tree display parameters are currently
left to the user to fine tune. A later version might automate some
of it, based on the size of the tree.

Since the stand-alone Ancestor Line Explorer is based on D3 v7, I
had to upgrade the other D3-based apps to v7 as well. The part I had
the most trouble with was zoom and pan. Documentation suggests one
should use .extent, .scaleExtent and .translateExtent to constrain
zoom and pan, but I could not get that to work so just did not use
it (other than .sclaeExtent). My testing did not show any adverse
effect.

All these changes can be test-driven at
https://apps.wikitree.com/apps/smit641/test2