Improve relative counts
Parent and spouse counts should now be correct for people "on the edge"
as well as for when a single degree is retrieved. Child and sibling
counts remain partial - only children and siblings present in the
CCn data retrieved are counted.

This change also includes highlighting (in the list of people) the
person for which a family group is being displayed