Improve relative counts
Relative count should now be correct.
In order to get the correct relative counts for each profile, we
get extra degrees (n+1 for CCn, and n-1, n, and n+1 if we get only
degree n). We make separate (async) calls for these in order to
be able to identify them (so as to not count and display them)
since currently the getPeople API does not return the degree
and calculating them is either not possible (in the single degree
case) or can lead to incorrect counts in the presence of private
profiles (in the CCn case).