CC7: Fix bug that allowed incorrect people to be shown in hiearchy
We get an extra degree of people in order to show correct relative
counts. The bug allowed the people in this extra degree to be shown
in the hierarchy and list views. However, their relative counts
are incorrect and therefore provide misleading "missing family"
information. These people are now no longer being shown.