CC7: Refactor and fix save/load problems
The cc7.js file was getting way too big, so it was split into
smaller files based on functionality. While testing the result
I noticed that save and load was not working correctly (nothing
to do with the refactor) so that was fixed, together with better
names for the saved files.