Update BioCheck messages that can be suppressed
The selection of messages that can be suppressed via the CC7 options
were updated to be in line with this commit:
https://github.com/ke4tch/wikitree-biocheck/commit/8aaca7437efb9ea036c3e4adf2d417787d01e18c
and this change on the BioCheck help page:
https://www.wikitree.com/index.php?title=Space:BioCheckHelp&diff=187073696&oldid=187062379

This change also includes a fix to the placement of the family sheet,
bio report and timeline tables when they are displayed for the second
time. Previously they were re-displayed at the same location as the
last time. However, this may be off the screen if the user had
applied a filter to the people table in the mean time. With this
change they are displayed at their initial position, close to the
button that allows them to be displayed.  The down-side is if the
user had dragged them to a different position for visibility reasons,
they will have to do so again. However, imho the surprise of not seeing
the pop-up (because it is off-screen) by far outweighs the annoyance
of having to drag them to a different place again.