CC7: correct behaviour around handling of private profiles
The API changed and no longer returns Name for private profiles.
It also does not return manager information if the manager's
profile is private (seemingly regardles of which "colour" of
private). The errors caused by this new behaviour has been fixed.
The down-side is that some profiles, for which one can see the
manager on WikiTree, are listed with 'Unknown' manager in CC7.