Make provision for the root profile not being returned by WT
A (presumably temporary) bug in WT is causing the requested
profile sometimes not to be returned together with its CCn
profiles. Previously this caused CC7 to throw an exception
and an unpleasant user experience. Now a warning is
displayed and some functionality that relies on the root
profile being present, is disabled.