CC7: Handle IDs that include spaces and are not capitalised
The TreeApps WT ID field may contain spaces (e.g. "St. Claire-3") and it is immaterial whether the first letter is capitalised or not, the profile will still be found. However, in the Name field being returned, the first letter is always capitalised and all spaces are replaces with underscores. This code ensures the CC7 Views app take that into account.

There is now a shared/Utils function that will return the entered field in its standard form.