Include timeline events from some templates
Previously we did this by scanning the bio, but since we don't always fetch the bio and if we do, we clear it immediately after we've done the bio check, the code had no effect. Now we request the templates via the api and will include war enlistings in the timeline.